feat: errors code for localization

+++ src/profile/profile.service.ts
@@ -1,4 +1,4 @@
-import { Injectable, NotFoundException, BadRequestException } from "@nestjs/common";
+import { Injectable, NotFoundException, ConflictException } from "@nestjs/common";
 import { InjectRepository } from "@nestjs/typeorm";
 import { Repository } from "typeorm";
 import { ProfileModel } from "./model/profile.model";
@@ -18,9 +18,10 @@ import {
   DeleteProfileInput,
 } from "src/graphql";
 
-const userNotFound = new NotFoundException("User not found");
-const languageHasBeenAdded = new BadRequestException("Language has already been added");
-const skillHasBeenAdded = new BadRequestException("Skill has already been added");
+const userNotFound = new NotFoundException("userNotFound");
+const languageHasBeenAdded = new ConflictException("languageHasBeenAdded");
+const skillHasBeenAdded = new ConflictException("skillHasBeenAdded");
+const profileNotFound = new NotFoundException("profileNotFound");
 
 @Injectable()
 export class ProfileService {
@@ -32,10 +33,16 @@ export class ProfileService {
     private readonly cloudService: CloudService,
   ) {}
 
-  findOneById(userId: string) {
-    return this.profileRepository.findOne({
+  async findOneById(userId: string) {
+    const profile = await this.profileRepository.findOne({
       where: { id: userId },
     });
+
+    if (!profile) {
+      throw profileNotFound;
+    }
+
+    return profile;
   }
 
   async me(userId: string) {
@@ -63,6 +70,11 @@ export class ProfileService {
 
   async updateProfile({ userId, first_name, last_name }: UpdateProfileInput) {
     const profile = await this.findOneById(userId);
+
+    if (!profile) {
+      throw profileNotFound;
+    }
+
     profile.first_name = first_name;
     profile.last_name = last_name;
     return await this.profileRepository.save(profile);
@@ -70,16 +82,28 @@ export class ProfileService {
 
   async addProfileSkill({ userId, name, categoryId, mastery }: AddProfileSkillInput) {
     const profile = await this.findOneById(userId);
+
+    if (!profile) {
+      throw profileNotFound;
+    }
+
     const index = profile.skills.findIndex((skill) => skill.name === name);
+
     if (index !== -1) {
       throw skillHasBeenAdded;
     }
+
     profile.skills.push({ name, categoryId, mastery });
     return await this.profileRepository.save(profile);
   }
 
   async updateProfileSkill({ userId, name, categoryId, mastery }: UpdateProfileSkillInput) {
     const profile = await this.findOneById(userId);
+
+    if (!profile) {
+      throw profileNotFound;
+    }
+
     const index = profile.skills.findIndex((skill) => skill.name === name);
 
     if (index !== -1) {
@@ -91,22 +115,39 @@ export class ProfileService {
 
   async deleteProfileSkill({ userId, name }: DeleteProfileSkillInput) {
     const profile = await this.findOneById(userId);
+
+    if (!profile) {
+      throw profileNotFound;
+    }
+
     profile.skills = profile.skills.filter((skill) => !name.includes(skill.name));
     return await this.profileRepository.save(profile);
   }
 
   async addProfileLanguage({ userId, name, proficiency }: AddProfileLanguageInput) {
     const profile = await this.findOneById(userId);
+
+    if (!profile) {
+      throw profileNotFound;
+    }
+
     const index = profile.languages.findIndex((skill) => skill.name === name);
+
     if (index !== -1) {
       throw languageHasBeenAdded;
     }
+
     profile.languages.push({ name, proficiency });
     return await this.profileRepository.save(profile);
   }
 
   async updateProfileLanguage({ userId, name, proficiency }: UpdateProfileLanguageInput) {
     const profile = await this.findOneById(userId);
+
+    if (!profile) {
+      throw profileNotFound;
+    }
+
     profile.languages = profile.languages.map((language) => {
       if (language.name === name) {
         return { name, proficiency };
@@ -118,6 +159,11 @@ export class ProfileService {
 
   async deleteProfileLanguage({ userId, name }: DeleteProfileLanguageInput) {
     const profile = await this.findOneById(userId);
+
+    if (!profile) {
+      throw profileNotFound;
+    }
+
     profile.languages = profile.languages.filter((language) => !name.includes(language.name));
     return await this.profileRepository.save(profile);
   }
@@ -132,12 +178,23 @@ export class ProfileService {
 
   async deleteAvatar({ userId }: DeleteAvatarInput) {
     const profile = await this.findOneById(userId);
+
+    if (!profile) {
+      throw profileNotFound;
+    }
+
     profile.avatar = null;
     await this.profileRepository.save(profile);
     return null;
   }
 
   async deleteProfile({ userId }: DeleteProfileInput) {
+    const profile = await this.findOneById(userId);
+
+    if (!profile) {
+      throw profileNotFound;
+    }
+
     return await this.profileRepository.delete(userId);
   }
 }
